System buffer-free version of ADC should now work.  Doing so, saves about 1kb of flash space.
NOTE: getsamples function is disabled since this is dependant on having a
buffer to get more than one sample.

Side note: single sampling without the buffer is faster by about 5-10
us/channel.  It should be possible to sidestep the buffer in certain
situations for buffer enabled mode and get a similar benefit at the cost of
some additional complexity.

git-svn-id: svn://svn.berlios.de/elua/trunk@267 67d3b963-1d53-0410-ba23-e466f4efc467