Added partial integration of AsciiDoc (http://www.methods.co.nz/asciidoc/) to our documentation system. For now:
- AsciiDoc doesn't work on automatically generated documentation (eluadoc/), just on the static content (but it will)
- Both AsciiDoc and the previous HTML versions are supported, HTML has priority over AsciiDoc

You need at least version 8.6.0 of AsciiDoc to test this.
To give an example of how cool AsciiDoc is, the News page (en/news.html) was rewritten in AsciiDoc format (en/news.txt). 
If you can't see the beauty of writing HTML in such a way that the overall structure of the document remains clear and easy to understand, I reserve the right to question your mental health :)
And yes, the long term goal is to have everything written in AsciiDoc.



git-svn-id: svn://svn.berlios.de/elua/trunk@794 67d3b963-1d53-0410-ba23-e466f4efc467